requirements.txt
asgiref==3.8.1
black==22.3.0
Django==5.1.3
djangorestframework==3.15.2
django-extensions==3.2.3
django-stubs==1.8.0
django-stubs-ext==0.3.1
factory_boy==3.3.3
flake8==7.0.0
isort==5.6.4
mypy==1.10.0
mypy-extensions==1.0.0
psycopg2-binary>=2.8
pytest>=7.2.1
pytest-django>=4.5.2
sqlparse==0.5.2
typing_extensions==4.12.2
